Calculate power at different points.
In this example, we use the NumPy library to create a loss array that contains the losses at each point in the laser system. We also define laser power as laser_power. Next, we create a power array of zeros with the same length as losses. We set the power at the first point to be equal to the laser power, and then we use a loop to calculate the power at each subsequent point. We do this by multiplying the previous power by (1 - losses[i-1]), which takes into account the losses at that point. Finally, we print out the total power at each point using a loop that iterates over the power array. The output will display the total power at each point in the laser system.