measure the spot size across various optics
If the file is in the same directory as your Python script, you can use the following code:
image = plt.imread('laser_beam.png')

However, if the file is located in a different directory, you can use an absolute or relative file path to access the file. Here are some examples:
# absolute path to the file
image = plt.imread('/path/to/laser_beam.png')

# relative path to the file
image = plt.imread('../images/laser_beam.png')

Make sure to replace "/path/to/" and "../images/" with the correct file path to the image file.