Create Spatial Intensity Profile 3D
We define a function beam_profile that returns the transverse profile of the beam, which in this case is a Gaussian profile. We also define a function intensity_profile that calculates the intensity profile of the beam along the propagation using the transverse profile and the beam radius at each point.

We then create a figure and axis for 3D plotting using the plt.figure() and plt.axes(projection='3d') functions.